manifest service worker PWA

+++ make-channel-list.js
@@ -18,17 +18,18 @@ request.get(apiUrl, (error, response, body) => {
 		variablesText='var channels='+JSON.stringify(channels)+',title='+JSON.stringify(title).replace(/\"([^(\")"]+)\":/g,"$1:")+',timeshift='+JSON.stringify(timeshift).replace(/\"([^(\")"]+)\":/g,"$1:")+',type='+JSON.stringify(type).replace(/\"([^(\")"]+)\":/g,"$1:");
 		// uncomment next line to console.log the result
 		//console.log("result:\n"+jsonText);
-		if(!fs.existsSync('channels.json')||jsonText!==fs.readFileSync('channels.json').toString()){
+		if(fs.existsSync('channels.js') && variablesText === fs.readFileSync('channels.js').toString()){console.log('No changes!');return;}
+		//if(!fs.existsSync('channels.json')||jsonText!==fs.readFileSync('channels.json').toString()){
 			fs.writeFile('channels.json', jsonText, (err, data)=>{
 				if(err)console.log("Unsuccessfully tried to write to channels.json file, error: ", err);
 				else console.log("Successfully written channels.json file");
-		});
-		}
-		if(!fs.existsSync('channels.js')||variablesText!==fs.readFileSync('channels.js').toString()){
+			});
+		//}
+		//if(!fs.existsSync('channels.js')||variablesText!==fs.readFileSync('channels.js').toString()){
 			fs.writeFile('channels.js', variablesText, (err, data)=>{
 				if(err)console.log("Unsuccessfully tried to write channels.js file, error: ", err);
 				else console.log("Successfully written channels.js file");
-		});
-		}
+			});
+		//}
 	});
 });
